requirements.txt
aiohttp
pandas
openpyxl
python-dotenv
supabase>=2.0.0
